os-prober: ensure test with cache works if os-prober is not installed
The test TestOsProber.test_run_once ensured that subprocess.run was
called only once when running the OS probe. It uses assert_called_once
for the assertion.

However, the _run_os_prober function would only call subprocess.run if
os-prober is detected/installed. When it is not, the call to subprocess
is skipped, which in turns result in the above assertion to be false.

To solve this, the lru_cache decoration is now moved to a new function
_cache_os_prober ; which calls _run_os_prober.

By ensuring that _run_os_prober is called only once (as opposed to
subprocess.run), the test case now works the same if os-prober is
or is not installed.

Signed-off-by: Olivier Gayot <olivier.gayot@sigexec.com>